:sparkles: hide during camera dragging :art: refactor
