+++ BuildScripts/_DoNotModify.txt
@@ -1,0 +1,1 @@
+﻿The BuildScripts here are used to package the module and copy over the resources included in the ref folder. Don't make changes unless you know what you are doing.
